fixed issue with resource type renaming
