+++ src/utils/nodeUtils.ts
@@ -9,3 +9,12 @@ export function isBranchState(state: State): boolean {
   const { action, label } = state as GuidanceState;
   return action === undefined && label !== 'Start';
 }
+
+type ConversionResource = {
+  [key: string]: string;
+};
+export const resourceNameConversion: ConversionResource = {
+  MedicationRequest: 'Medication',
+  ServiceRequest: 'Procedure',
+  Careplan: 'Regimen'
+};
